refactor: 프로젝트명 변경 및 코드 스타일 통일
