+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'blok'
+rootProject.name = 'took'
