--- src/main/java/com/evenly/blok/BlokApplication.java
@@ -1,13 +1,0 @@
-package com.evenly.blok;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BlokApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BlokApplication.class, args);
-	}
-
-}
