--- src/main/java/com/evenly/blok/global/exception/BlokException.java
@@ -1,11 +1,0 @@
-package com.evenly.blok.global.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class BlokException extends RuntimeException { // TODO: 서비스명에 맞게 변경
-
-    private final ErrorCode errorCode;
-}
