--- src/main/java/com/evenly/blok/global/exception/ErrorCode.java
@@ -1,10 +1,0 @@
-package com.evenly.blok.global.exception;
-
-import org.springframework.http.HttpStatus;
-
-public interface ErrorCode {
-
-    HttpStatus getStatus();
-
-    String getMessage();
-}
