+++ src/main/java/com/evenly/took/TookApplication.java
@@ -1,0 +1,13 @@
+package com.evenly.took;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TookApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TookApplication.class, args);
+	}
+
+}
