+++ src/main/java/com/evenly/took/global/common/constants/UrlConstants.java
@@ -1,0 +1,14 @@
+package com.evenly.took.global.common.constants;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum UrlConstants {
+
+	LOCAL_SERVER_URL("http://localhost:8080"),
+	;
+
+	private final String value;
+}
