+++ src/main/java/com/evenly/took/global/exception/ErrorCode.java
@@ -1,0 +1,10 @@
+package com.evenly.took.global.exception;
+
+import org.springframework.http.HttpStatus;
+
+public interface ErrorCode {
+
+	HttpStatus getStatus();
+
+	String getMessage();
+}
