+++ src/main/java/com/evenly/took/global/exception/TookException.java
@@ -1,0 +1,11 @@
+package com.evenly.took.global.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class TookException extends RuntimeException {
+
+	private final ErrorCode errorCode;
+}
