--- src/test/java/com/evenly/blok/BlokApplicationTests.java
@@ -1,15 +1,0 @@
-package com.evenly.blok;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-
-@SpringBootTest
-@ActiveProfiles("test")
-class BlokApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
