--- src/test/java/com/evenly/blok/global/config/TestcontainersConfig.java
@@ -1,21 +1,0 @@
-package com.evenly.blok.global.config;
-
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
-import org.springframework.context.annotation.Bean;
-import org.testcontainers.containers.MySQLContainer;
-import org.testcontainers.utility.DockerImageName;
-
-@TestConfiguration(proxyBeanMethods = false)
-public class TestcontainersConfig {
-
-	@Bean
-	@ServiceConnection
-	MySQLContainer<?> mysqlContainer() {
-		return new MySQLContainer<>(DockerImageName.parse("mysql/mysql-server:8.0.26"))
-			.withDatabaseName("testdb")
-			.withUsername("test")
-			.withPassword("test");
-	}
-
-}
