--- src/test/java/com/evenly/blok/global/health/HealthControllerIntegrationTest.java
@@ -1,20 +1,0 @@
-package com.evenly.blok.global.health;
-
-import static io.restassured.RestAssured.*;
-
-import org.junit.jupiter.api.Test;
-
-import com.evenly.blok.global.integration.IntegrationTest;
-
-class HealthControllerIntegrationTest extends IntegrationTest {
-
-	@Test
-	void Health_Check_통합_테스트_성공() {
-		given().log().all()
-			.when()
-			.get("/api/health")
-			.then()
-			.log().all()
-			.statusCode(200);
-	}
-}
