--- src/test/java/com/evenly/blok/global/repository/RepositoryTest.java
@@ -1,18 +1,0 @@
-package com.evenly.blok.global.repository;
-
-import static com.evenly.blok.global.common.constants.EnvironmentConstants.*;
-
-import org.junit.Ignore;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-@DataJpaTest
-@ExtendWith(SpringExtension.class)
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-@ActiveProfiles(TEST_ENV)
-@Ignore
-public abstract class RepositoryTest {
-}
