--- src/test/java/com/evenly/blok/global/service/MockTest.java
@@ -1,14 +1,0 @@
-package com.evenly.blok.global.service;
-
-import static com.evenly.blok.global.common.constants.EnvironmentConstants.*;
-
-import org.junit.Ignore;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.test.context.ActiveProfiles;
-
-@ExtendWith(MockitoExtension.class)
-@ActiveProfiles(TEST_ENV)
-@Ignore
-public abstract class MockTest {
-}
