+++ src/test/java/com/evenly/took/TookApplicationTests.java
@@ -1,0 +1,15 @@
+package com.evenly.took;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+@SpringBootTest
+@ActiveProfiles("test")
+class TookApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
