+++ src/test/java/com/evenly/took/global/integration/IntegrationTest.java
@@ -1,0 +1,37 @@
+package com.evenly.took.global.integration;
+
+import static com.evenly.took.global.common.constants.EnvironmentConstants.*;
+
+import org.junit.Ignore;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import io.restassured.RestAssured;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@ExtendWith(SpringExtension.class)
+@Transactional
+@ActiveProfiles(TEST_ENV)
+@Ignore
+public abstract class IntegrationTest {
+
+	@LocalServerPort
+	protected int port;
+
+	@Autowired
+	protected ObjectMapper objectMapper;
+
+	@BeforeEach
+	public void setUp() {
+		RestAssured.baseURI = "http://localhost";
+		RestAssured.port = port;
+	}
+}
