+++ src/test/java/com/evenly/took/global/repository/RepositoryTest.java
@@ -1,0 +1,18 @@
+package com.evenly.took.global.repository;
+
+import static com.evenly.took.global.common.constants.EnvironmentConstants.*;
+
+import org.junit.Ignore;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+@DataJpaTest
+@ExtendWith(SpringExtension.class)
+@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
+@ActiveProfiles(TEST_ENV)
+@Ignore
+public abstract class RepositoryTest {
+}
